requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cloudinary-storage==0.3.0
drf-yasg==1.20.0
django-redis==5.2.0
dj-database-url==0.5.0
django-cors-headers==3.13.0
gunicorn==20.1.0
whitenoise==6.2.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pillow==9.1.1
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
pyjwt==2.4.0
redis==4.3.3
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
